updated CTC decoding and added comments
