+++ config/libri/ctc_decode_example.yaml
@@ -1,0 +1,16 @@
+# Most of the parameters will be imported from the training config
+src:
+  ckpt: 'ckpt/asr_example_sd0/best_ctc.pth'
+  config: 'config/libri/asr_example.yaml'
+data:
+  corpus:
+    name:  'Librispeech'
+    dev_split: ['dev-clean']
+    test_split: ['test-clean']
+decode:
+  beam_size: 20
+  vocab_candidate: 30
+  lm_path: 'ckpt/lm_example_sd0/best_ppx.pth'
+  lm_config: 'config/libri/lm_example.yaml'
+  lm_weight: 0.5
+  ctc_weight: 1.0
